# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "Stable, for your code",
                "type": "string"
              },
              "details": {
                "description": "Depends on code",
                "type": "object"
              },
              "message": {
                "description": "For people; may change",
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "QueueBorderlineOut": {
        "properties": {
          "evaluation_id": {
            "title": "Evaluation Id",
            "type": "string"
          },
          "tier_above": {
            "title": "Tier Above",
            "type": "string"
          },
          "tier_below": {
            "title": "Tier Below",
            "type": "string"
          }
        },
        "required": [
          "evaluation_id",
          "tier_above",
          "tier_below"
        ],
        "title": "QueueBorderlineOut",
        "type": "object"
      },
      "QueueItemOut": {
        "properties": {
          "application_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id"
          },
          "borderline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueueBorderlineOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          },
          "resolve_at": {
            "title": "Resolve At",
            "type": "string"
          },
          "waiting_since": {
            "title": "Waiting Since",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "candidate_id",
          "application_id",
          "reason_code",
          "reason",
          "borderline",
          "waiting_since",
          "resolve_at"
        ],
        "title": "QueueItemOut",
        "type": "object"
      },
      "QueueKindCountOut": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "oldest_since": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Since"
          },
          "oldest_waiting_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Waiting Hours"
          },
          "open": {
            "title": "Open",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "open",
          "oldest_since",
          "oldest_waiting_hours"
        ],
        "title": "QueueKindCountOut",
        "type": "object"
      },
      "QueuePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "QueuePage",
        "type": "object"
      },
      "QueueSummaryOut": {
        "properties": {
          "kinds": {
            "items": {
              "$ref": "#/components/schemas/QueueKindCountOut"
            },
            "title": "Kinds",
            "type": "array"
          },
          "oldest_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Kind"
          },
          "oldest_waiting_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Waiting Hours"
          },
          "open": {
            "title": "Open",
            "type": "integer"
          }
        },
        "required": [
          "open",
          "oldest_waiting_hours",
          "oldest_kind",
          "kinds"
        ],
        "title": "QueueSummaryOut",
        "type": "object"
      },
...
      "ResolutionOut": {
        "properties": {
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "resolved_at": {
            "title": "Resolved At",
            "type": "string"
          },
          "resolved_by": {
            "title": "Resolved By",
            "type": "string"
          },
          "transition_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transition Id"
          }
        },
        "required": [
          "decision",
          "reason",
          "transition_id",
          "resolved_by",
          "resolved_at"
        ],
        "title": "ResolutionOut",
        "type": "object"
      },
...
      "ReviewItemOut": {
        "properties": {
          "application_id": {
            "title": "Application Id",
            "type": "string"
          },
          "at_stage": {
            "title": "At Stage",
            "type": "string"
          },
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "proposed_at": {
            "title": "Proposed At",
            "type": "string"
          },
          "proposed_by": {
            "title": "Proposed By",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          },
          "requisition_id": {
            "title": "Requisition Id",
            "type": "string"
          },
          "resolution": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "kind",
          "application_id",
          "candidate_id",
          "requisition_id",
          "at_stage",
          "reason_code",
          "reason",
          "proposed_by",
          "proposed_at",
          "resolution"
        ],
        "title": "ReviewItemOut",
        "type": "object"
      },
      "ReviewItemPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ReviewItemPage",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "A sign-in token from the company identity provider, or from /dev/token in dev.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Talent Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/reports/review-queue": {
      "get": {
        "description": "Open items per kind, and how long the oldest has waited. Counts only.",
        "operationId": "review_queue_report_v1_reports_review_queue_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueSummaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Queue Report",
        "tags": [
          "review queue",
          "reports"
        ]
      }
    },
...
    "/v1/review-items": {
      "get": {
        "description": "Items in your scope, newest first. Open items by default.",
        "operationId": "list_review_items_v1_review_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "open",
              "enum": [
                "open",
                "resolved"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "negative_verdict",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "requisition_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requisition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItemPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Review Items",
        "tags": [
          "review queue"
        ]
      }
    },
...
    "/v1/review-queue": {
      "get": {
        "description": "Every open item in your scope, of every kind, oldest first, each with its reason.",
        "operationId": "review_queue_v1_review_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "negative_verdict",
                    "flagged_document",
                    "unverified_candidate",
                    "possible_duplicate",
                    "borderline_score",
                    "ai_assessment"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Queue",
        "tags": [
          "review queue"
        ]
      }
    },
...
  }
}